test(e2e): Group duplicated code on AP tests
